refactor(components): extract non-component exports for fast refresh
react-refresh/only-export-components requires component files to export
only components. Moved, verbatim:

- MemberNode.tsx ring/avatar helpers (getRingGradient, getRingShadow,
  getFallbackGradient, getInitials) -> new components/memberVisuals.tsx.
  PersonAvatarIcon stays in MemberNode.tsx — it IS a component, and
  placing it beside the helpers would just move the violation into the
  new file.
- AdvancedFilter.tsx filter model (FilterState, DEFAULT_FILTERS,
  isDefaultFilter) -> new views/filterState.ts, with a type-only
  re-export so existing `import type { FilterState }` sites keep
  working untouched.

Import sites updated mechanically; no logic changed.
