+++ src/components/admin/AdminDashboard.tsx
@@ -6,7 +6,8 @@ import { useFamilyStore } from '../../store/useFamilyStore'
 import { useLang, isRTL, type Translations } from '../../i18n/useT'
 import EditMemberModal from '../EditMemberModal'
 import InviteCodeManager from './InviteCodeManager'
-import { getRingGradient, getFallbackGradient, PersonAvatarIcon } from '../MemberNode'
+import { PersonAvatarIcon } from '../MemberNode'
+import { getRingGradient, getFallbackGradient } from '../memberVisuals'
 import type { EditRequest, Member, Profile, UserRole, MasterPermissions, AccessRequest } from '../../types'
 import type { PermissionKey } from '../../lib/permissions'
