+++ src/pages/Dashboard.tsx
@@ -5,7 +5,8 @@ import { useFamilyStore } from '../store/useFamilyStore'
 import { useLang, isRTL } from '../i18n/useT'
 import { supabase } from '../lib/supabase'
 import { isAdmin, isOnboarded } from '../lib/permissions'
-import { getRingGradient, getFallbackGradient, PersonAvatarIcon } from '../components/MemberNode'
+import { PersonAvatarIcon } from '../components/MemberNode'
+import { getRingGradient, getFallbackGradient } from '../components/memberVisuals'
 import AIScanModal from '../components/ai/AIScanModal'
 import ComingSoonModal from '../components/ComingSoonModal'
 import BuildFromTextModal from '../components/BuildFromTextModal'
