+++ src/pages/TreePage.tsx
@@ -12,7 +12,8 @@ import TreeSearchModal from '../components/TreeSearchModal'
 import TreeSwitcher from '../components/TreeSwitcher'
 import EditModeToggle from '../components/EditModeToggle'
 import { useEffect, useMemo, useState } from 'react'
-import AdvancedFilter, { DEFAULT_FILTERS, type FilterState } from '../components/views/AdvancedFilter'
+import AdvancedFilter from '../components/views/AdvancedFilter'
+import { DEFAULT_FILTERS, type FilterState } from '../components/views/filterState'
 import { useBrowserZoom } from '../hooks/useBrowserZoom'
 import { useHorizontalSwipe } from '../hooks/useHorizontalSwipe'
 import Tooltip from '../components/Tooltip'
